Move fleet var funcs
